+++ cmd/pmem-vgm/main_test.go
@@ -3,9 +3,9 @@ package main_test
 import (
 	"testing"
 
-	"github.com/intel/pmem-csi/pkg/pmem-ns-init"
+	"github.com/intel/pmem-csi/pkg/pmem-vgm"
 )
 
 func TestMain(t *testing.T) {
-	pmemnsinit.Main()
+	pmemvgm.Main()
 }
