+++ deploy/kubernetes-1.13/pmem-csi-lvm.yaml
@@ -271,11 +271,15 @@ spec:
         - -namespacesize=9
         command:
         - /go/bin/pmem-ns-init
+        env:
+        - name: TERMINATION_LOG_PATH
+          value: /tmp/pmem-ns-init-termination-log
         image: 192.168.8.1:5000/pmem-csi-driver:canary
         imagePullPolicy: Always
         name: pmem-ns-init
         securityContext:
           privileged: true
+        terminationMessagePath: /tmp/pmem-ns-init-termination-log
         volumeMounts:
         - mountPath: /sys
           name: sys-dir
@@ -283,11 +287,15 @@ spec:
         - -v=3
         command:
         - /go/bin/pmem-vgm
+        env:
+        - name: TERMINATION_LOG_PATH
+          value: /tmp/pmem-vgm-termination-log
         image: 192.168.8.1:5000/pmem-csi-driver:canary
         imagePullPolicy: Always
         name: pmem-vgm
         securityContext:
           privileged: true
+        terminationMessagePath: /tmp/pmem-vgm-termination-log
       nodeSelector:
         storage: pmem
       volumes:
